move AbstractDataBindingAdapter to gto-support-androidx-recyclerview
