+++ gto-support-androidx-recyclerview/build.gradle.kts
@@ -7,6 +7,7 @@ configureAndroidLibrary()
 
 dependencies {
     api(libs.androidx.recyclerview)
+    implementation(libs.weakdelegate)
 
     // region *DataBindingAdapter
     compileOnly(libs.androidx.databinding.runtime)
